Leave range requests up to nginx
This along with the previous couple of test commits, fixes how we send
catalog files to the browser.

We use the standard X-Sendfile functionality built into send_file so
that rails tells the web server (nginx) in our case which file to send
to the browser.

This means that the rails process isn't held up streaming files to the
browser and can process more requests. This solves a problem with the
way the catalog works where it might hang for a minute while files are
pulled from tape.

It has the added benefit of not needing to have our own range request
logic as nginx will handle that for us.